requirements.txt
asgiref==3.5.2
boto3==1.17.52
botocore==1.20.52
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-ckeditor==5.9.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-import-export==2.8.0
django-js-asset==1.2.2
django-markdownx==3.0.1
django-sendgrid-v5==0.8.1
django-storages==1.11.1
djangorestframework==3.11.0
et-xmlfile==1.1.0
future==0.18.2
idna==2.9
jmespath==0.10.0
Markdown==3.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow
psycopg2-binary
python-dateutil==2.8.1
python-http-client==3.2.4
pytz==2019.3
PyYAML==6.0
razorpay==1.2.0
requests==2.23.0
s3transfer==0.3.7
sendgrid==6.1.1
six==1.15.0
sqlparse==0.3.0
tablib==3.2.1
urllib3==1.25.8
xlrd==2.0.1
xlwt==1.3.0
